backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
google-generativeai==0.3.2
joblib==1.3.2